requirements.txt
anyio==4.3.0
bottle==0.12.25
certifi==2024.2.2
cffi==1.16.0
clr-loader==0.2.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ollama==0.1.9
proxy_tools==0.1.0
pycparser==2.22
pythonnet==3.0.3
pywebview==5.1
sniffio==1.3.1
typing_extensions==4.11.0
